fixed some softfloat warnings and removed frame buffer and native page types, they will now map their memory using the normal rw ram page and use the new ram indexing to map the native ram to emulated ram.

+++ lib/softfloat/source/extF80_div.c
@@ -173,7 +173,7 @@ extFloat80_t extF80_div( extFloat80_t a, extFloat80_t b )
     *------------------------------------------------------------------------*/
  propagateNaN:
     uiZ = softfloat_propagateNaNExtF80UI( uiA64, uiA0, uiB64, uiB0 );
-    uiZ64 = uiZ.v64;
+    uiZ64 = (uint_fast16_t)uiZ.v64;
     uiZ0  = uiZ.v0;
     goto uiZ;
     /*------------------------------------------------------------------------
